test: extract fixtures to a separate file
