+++ src/modules/private-share-folder/private-sharing.usecase.spec.ts
@@ -31,6 +31,7 @@ import { ConfigModule } from '@nestjs/config';
 import configuration from '../../config/configuration';
 import { ForbiddenException } from '@nestjs/common';
 import { PrivateSharingFolderRolesRepository } from './private-sharing-folder-roles.repository';
+import { newFolder, newUser } from '../../../test/fixtures';
 
 describe('Private Sharing Use Cases', () => {
   let privateSharingUseCase: PrivateSharingUseCase;
@@ -1215,54 +1216,4 @@ describe('Private Sharing Use Cases', () => {
       expect(removeSharedWithMock).toHaveBeenCalled();
     });
   });
-
-  function newUser(): User {
-    return User.build({
-      id: Math.random() * 999999,
-      userId: '',
-      name: 'John',
-      lastname: 'Doe',
-      uuid: v4(),
-      email: '',
-      username: '',
-      bridgeUser: '',
-      password: '',
-      mnemonic: '',
-      referrer: v4(),
-      referralCode: v4(),
-      credit: 0,
-      hKey: Buffer.from(''),
-      rootFolderId: 0,
-      errorLoginCount: 0,
-      isEmailActivitySended: 0,
-      lastResend: new Date(),
-      syncDate: new Date(),
-      welcomePack: false,
-      registerCompleted: false,
-      secret_2FA: '',
-      backupsBucket: '',
-      sharedWorkspace: false,
-      tempKey: '',
-      avatar: '',
-    });
-  }
-
-  function newFolder(owner?: User): Folder {
-    return Folder.build({
-      id: Math.random(),
-      uuid: v4(),
-      name: 'folder',
-      parentId: 0,
-      userId: owner?.id ?? 0,
-      createdAt: new Date(),
-      updatedAt: new Date(),
-      bucket: '',
-      plainName: '',
-      encryptVersion: '03-aes',
-      deleted: false,
-      removed: false,
-      deletedAt: undefined,
-      removedAt: undefined,
-    });
-  }
 });
